+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.10.0-DEV] - 2021-03-08
+- added DcaGenerator class
+- added EntrypointCollectionFactory and EntrypointCollection class
+- added TemplateAssetGenerator class
+
 ## [1.9.2] - 2021-03-03
 - fixed ordering issue with use_contao_template_variables
 - added notice to layout settings if use_contao_template_variables used together with "Add jQuery"